refactor: memberPage 마이그레이션
