+++ src/components/Common/SectionHeader/SectionHeader.tsx
@@ -13,6 +13,7 @@ interface SectionHeaderProps {
 const SectionHeader = ({ name, link, state }: SectionHeaderProps) => {
   return (
     <div className={container}>
+      {/* Heading 컴포넌트로 교체 weight bold */}
       <h1 className={title}>{name}</h1>
       {link && (
         <Link to={link} state={state}>
