+++ src/components/Common/SectionHeader/sectionHeader.css.ts
@@ -3,7 +3,6 @@ import { style } from '@vanilla-extract/css';
 export const container = style({
   display: 'flex',
   justifyContent: 'space-between',
-  padding: '0 20px',
 });
 
 export const title = style({
