+++ src/components/Common/index.ts
@@ -25,3 +25,4 @@ export { default as PageHeader } from './PageHeader/PageHeader';
 export { default as Badge } from './Badge/Badge';
 export { default as WriteButton } from './WriteButton/WriteButton';
 export { default as Text } from './Text/Text';
+export { default as TopBar } from './TopBar/TopBar';
